opt: rearrange the matrix and map of connections
